requirements.txt
gradio
onnxruntime
pillow
numpy
torch